Use tftpboot instead of using image written by TI flash tool
I've noticed issues on at least one SOM, with MAC 00:08:33:40:A1:50,
where the TI flash tool programming sequence was not able to
successfully bring the system to a running state.  The problem appears
to be that the data written to the ubifs was not read successfully with
1-bit ECC supported by the flash tool.

To resolve, avoid writing the data with the TI flash tool and instead
read as much as possible using tftpboot.  This includes everything other
than the mlo and u-boot used for bootstrap.